adapting app code for better QT understanding and fix of data management, mostyl for aux file but for general usage too
